refactor, add complex test for traversal (#427)
